requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-docx==0.8.11
PyPDF2==3.0.1
openai==0.28.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
google-generativeai==0.3.2
